add resend function to varification email - make tokens expire after half-an-hour - fix cute_image out of index error - emails can be resent if user needs!

+++ accounts/views.py
@@ -1,3 +1,4 @@
+from django.utils import timezone
 from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
 from django.contrib import auth
@@ -6,7 +7,7 @@ from django.contrib.sites.shortcuts import get_current_site
 from django.http import HttpResponse, HttpResponseRedirect
 from . import pyfunctions
 from .models import Temp_user
-
+from django.urls import reverse
 
 def login(request):
     if request.user.is_authenticated:
@@ -66,14 +67,13 @@ def signup(request):
                 Temp_user.objects.create(uname=request.POST['username'], password=make_password(request.POST['password1']), email=request.POST['email'], token=token)
                 reciever = [request.POST["email"]]
                 if pyfunctions.varification_mailto(reciever, token):
-                    responce = ["Email Sent", "Please Varify you email!"]
+                    img = '<img src="' + pyfunctions.get_cute_image() + '" height="200px" width="200px" alt="a cute animal image">'
+                    return render(request, 'accounts/resend.html', {'title':'email varifiaction','case':'first_time', 'token':token, 'cute_image': img})
                 else:
                     responce = ["ERROR! mail not sent"] 
-                
                 img = pyfunctions.get_cute_image()
-                responce.append('<img src="' + img + '" height=50% width=50% alt="">')
-                return render(request, 'accounts/message.html', {'messages':responce})
-            
+                responce.append('<img src="' + img + '" height="200px" width="200px" alt="a cute animal image">')
+                return render(request, 'accounts/message.html', {'title': 'signed up', 'messages':responce})
             if error:
                 return render(request, 'accounts/signup.html', {'title':'signup error', 'error':error})
         
@@ -89,23 +89,52 @@ def activate(request):
 
     if Temp_user.objects.filter(token=key).exists():
         tuser = Temp_user.objects.get(token=key)
-        # Adding to user
-        user = User.objects.create_user(username=tuser.uname, password=tuser.password, email=tuser.email)
-        user.password = tuser.password # Using this since django will re-hash the hashed password => Thus explicitely, re mentioning it!
-        user.is_active = True
-        user.save()
+        
+        if tuser.varify_time(timezone.now()): # checking whether the link is valid!
+            # Adding to user
+            user = User.objects.create_user(username=tuser.uname, password=tuser.password, email=tuser.email)
+            user.password = tuser.password # Using this since django will re-hash the hashed password => Thus explicitely, re mentioning it!
+            user.is_active = True
+            user.save()
 
-        # Deleting from Temp_user
-        Temp_user.objects.filter(uname=tuser.uname).delete()
-        message = [
-            "Welcome to snapshare!", 
-            """Thanks for varifying your email! try <a href="{% url 'login' %}">logging in!"""
-        ]
+            # Deleting from Temp_user
+            Temp_user.objects.filter(uname=tuser.uname).delete()
+            message = [
+                "Welcome to snapshare!", 
+                """Thanks for varifying your email! try <a href="{% url 'login' %}">logging in! </a>"""
+            ]
+        
+        else:
+            img = '<img src="' + pyfunctions.get_cute_image() + '" height="200px" width="200px" alt="a cute animal image">'
+            return render(request, 'accounts/resend.html', {'title':"email error",'case':'expire', 'token':tuser.token, 'cute_image': img})
     
     else:
         title = "404"
         message = ["This way does not leads to mars!"]
 
     img = pyfunctions.get_cute_image()
-    message.append('<img src="' + img + '" height="200px" width="200px" alt="">')
+    message.append('<img src="' + img + '" height="200px" width="200px" alt="a cute animal image">')
     return render(request, 'accounts/message.html', {'title':title, 'messages':message})
+
+
+def resend(request):
+    if request.method == 'POST' and Temp_user.objects.filter(token=request.POST["token"]).exists():
+        token = request.POST['token']
+        tuser = Temp_user.objects.get(token=token)
+        new_token = pyfunctions.generate_url()
+        tuser.token = new_token
+        tuser.save()
+        reciver = [tuser.email]
+        token = new_token
+        if pyfunctions.varification_mailto(reciver, token):
+            img = '<img src="' + pyfunctions.get_cute_image() + '" height="200px" width="200px" alt="a cute animal image">'
+            return render(request, 'accounts/resend.html', {'title':'email varifiaction','case':'resend', 'token':new_token, 'cute_image': img})
+        else:
+            message = ["ERROR! mail not sent"] 
+    else:
+        title = "404"
+        message = ["Well, you have came a long way, to vain!"]
+    message.append('<img src="' + pyfunctions.get_cute_image() + '" height="200px" width="200px" alt="a cute animal image">')
+    return render(request, 'accounts/message.html', {'title':"404", 'messages':message})
+
+
